Opus021

+++ src/la/clamor/Instrument.java
@@ -10,8 +10,8 @@ package la.clamor;
  * @author mina
  */
 public interface Instrument {
-    public Legibilis capioNotum(int note, double diuturnitas, Velocitas velocitas);
-    public Legibilis capioNotum(int octave, int note, double diuturnitas, Velocitas velocitas);
+    public Legibilis capioNotum(double note, double diuturnitas, Velocitas velocitas);
+    public Legibilis capioNotum(int octave, double note, double diuturnitas, Velocitas velocitas);
     
     public String getName();
 }
